维权机器人/requirements.txt
python-telegram-bot==20.7
asyncpg
python-dotenv
reportlab
pillow
